+++ src/frontend/packages/lib_video/src/components/common/Player/videojs/downloadVideoPlugin/types.ts
@@ -1,0 +1,12 @@
+import { urls } from 'lib-components';
+import videojs from 'video.js';
+
+export interface DownloadVideoQualityItemOptions
+  extends videojs.MenuItemOptions {
+  label: string;
+  src: string | undefined;
+}
+
+export interface DownloadVideoPluginOptions {
+  urls: Partial<urls>;
+}