+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add a command to delete expired classrooms / videos
 - Add retention duration on playlist form
 - Add retention date widget to video and classroom
+- Add a download-video widget directly in the video player
 
 ### Changed